Fix: Reset posts page to 1 when clicking on a post
Previously, there was a bug where clicking on a post didn't reset the posts page to 1. This fix ensures that when a user clicks on a post, the posts page is now correctly set back to 1. This resolves the issue where pressing the home button didn't return the user to the first page of posts.